+++ node/commons/inputs.ts
@@ -1,22 +1,14 @@
-export interface TopSearchesInput {
-  store: string;
-  paidNavigation?: boolean;
-}
-
 export interface SuggestionSearchesInput {
-  store: string;
   term: string;
 }
 
 export interface SuggestionProductsInput {
-  store: string;
   term: string;
   attributeKey?: string;
   attributeValue?: string;
 }
 
 export interface SearchResultInput {
-  store: string;
   attributePath: string;
   query: string;
   page: number;