+++ react/index.ts
@@ -1,7 +1,5 @@
-import { Product } from "./models/product";
 import BiggyClient from "./utils/biggy-client";
 
 export default {
   BiggyClient,
-  Product,
 };